Update contacts.js and patient.js

+++ routes/patient.js
@@ -1,9 +1,6 @@
 const express = require("express");
 const async = require("hbs/lib/async");
 const { DATETIME2 } = require("mysql/lib/protocol/constants/types");
-//const { DATETIME, DATETIME2 } = require("mysql/lib/protocol/constants/types");
-//const async = require("hbs/lib/async");
-//const { JSON } = require("mysql/lib/protocol/constants/types");
 const router = express.Router();
 const pool = require("../db/db");
 
@@ -18,8 +15,7 @@ router.get("/", (req, res) => {
     
     if (!err) {
         res.send(result)
-     // res.json(rows);
-      console.log(fiels);
+        console.log(fiels);
     } else {
       console.log(err);
     }
@@ -27,15 +23,9 @@ router.get("/", (req, res) => {
 });
 router.post("/", async (req, res) => {
      var patient= req.body
-    // console.log(patient)
-   //  med=JSON.stringify(patient.medications)
-    // all=JSON.stringify(patient.allergies)
-   //  sur=JSON.stringify(patient.surgeries)
-   //  que=JSON.stringify(patient.questionnaire)
-   //  sym=JSON.stringify(patient.symptoms)
-   //  fam=JSON.stringify(patient.familyHistory)
-   //  add=JSON.stringify(patient.addictions)
-    result = await validatePatient(patient)
+      console.log(patient.length)
+    
+      result = await validatePatient(patient)
     if ( result=="ok"){
      med=JSON.stringify(patient.medications)
      all=JSON.stringify(patient.allergies)
@@ -44,7 +34,6 @@ router.post("/", async (req, res) => {
      sym=JSON.stringify(patient.symptoms)
      fam=JSON.stringify(patient.familyHistory)
      add=JSON.stringify(patient.addictions)
-   //   pool.query("INSERT INTO Patient (modifiedAt,email,firstName,lastName,phone,ssn,dateOfBirth,street,city,state,zip,insuranceCompany,plan,groupNumber,medications,surgeries,familyHistory,addictions,questionnaire,signature,middleInitial) VALUES ('2007-05-08 12:35:29.123','"+patient.email+"','"+patient.firstName+"','"+patient.lastName+"','"+patient.phone+"','"+patient.ssn+"','"+patient.dateOfBirth+"','"+patient.street+"','"+patient.city+"','"+patient.state+"','"+patient.zip+"','"+patient.insuranceCompany+"','"+patient.plan+"','"+patient.groupNumber+"','"+patient.medications+"','"+patient.surgeries+"','"+patient.familyHistory+"','"+patient.addictions+"','"+patient.questionnaire+"','"+patient.signature+"','"+patient.middleInitial+"');" ,(err, rows, fiels) => {  
        pool.query("INSERT INTO Patients (title,firstName,middleInitial,lastName,phone,email,sex,ssn,dateOfBirth,street,city,state,zip,insuranceCompany,plan,groupNumber,cardholder,medications,allergies,surgeries,familyHistory,addictions,questionnaire,symptoms,signature) VALUES ('"+patient.title+"','"+patient.firstName+"','"+patient.middleInitial+"','"+patient.lastName+"','"+patient.phone+"','"+patient.email+"','"+patient.sex+"','"+patient.ssn+"','"+patient.dateOfBirth+"','"+patient.street+"','"+patient.city+"','"+patient.state+"','"+patient.zip+"','"+patient.insuranceCompany+"','"+patient.plan+"','"+patient.groupNumber+"','"+patient.cardHolder+"','"+med+"','"+all+"','"+sur+"','"+fam+"','"+add+"','"+que+"','"+sym+"','"+patient.signature+"');" ,(err, rows, fiels) => {  
     if (!err) {
       
@@ -55,7 +44,6 @@ router.post("/", async (req, res) => {
       console.log(err);
     }
   });
-   //  res.status(201).send("Patient "+patient.firstName+" "+patient.lastName+" has beed added to the patient list.")
     }
      else{
       res.status(400).send(result)
@@ -64,6 +52,7 @@ router.post("/", async (req, res) => {
 
 router.put("/", async  (req, res) => {
     var patient= req.body
+    
      result= await validatePatient(patient);
     if (result=="ok"){
      med=JSON.stringify(patient.medications)
@@ -73,10 +62,7 @@ router.put("/", async  (req, res) => {
      sym=JSON.stringify(patient.symptoms)
      fam=JSON.stringify(patient.familyHistory)
      add=JSON.stringify(patient.addictions)
-   //   pool.query("INSERT INTO Patient (modifiedAt,email,firstName,lastName,phone,ssn,dateOfBirth,street,city,state,zip,insuranceCompany,plan,groupNumber,medications,surgeries,familyHistory,addictions,questionnaire,signature,middleInitial) VALUES ('2007-05-08 12:35:29.123','"+patient.email+"','"+patient.firstName+"','"+patient.lastName+"','"+patient.phone+"','"+patient.ssn+"','"+patient.dateOfBirth+"','"+patient.street+"','"+patient.city+"','"+patient.state+"','"+patient.zip+"','"+patient.insuranceCompany+"','"+patient.plan+"','"+patient.groupNumber+"','"+patient.medications+"','"+patient.surgeries+"','"+patient.familyHistory+"','"+patient.addictions+"','"+patient.questionnaire+"','"+patient.signature+"','"+patient.middleInitial+"');" ,(err, rows, fiels) => {  
-         //pool.query("UPDATE Patients SET title='"+patient.title+"',firstName='"+patient.firstName+"',middleInitial='"+patient.middleInitial+"',lastName='"+patient.lastName+"',phone='"+patient.phone+"',email='"+patient.email+"',sex='"+patient.sex+"',ssn='"+patient.ssn+"',dateOfBirth='"+patient.dateOfBirth+"',street='"+patient.street+"',city='"+patient.city+"',state='"+patient.state+"',zip='"+patient.zip+"',insuranceCompany='"+patient.insuranceCompany+"',plan='"+patient.plan+"',groupNumber='"+patient.groupNumber+"',cardHolder='"+patient.cardHolder+"',medications='"+patient.medications+"',surgeries='"+patient.surgeries+"',familyHistory='"+patient.familyHistory+"',addictions='"+patient.addictions+"',questionnaire='"+patient.questionnaire+"',symptoms='"+patient.symptoms+"',signature='"+patient.signature+"' WHERE id = '"+patient.id+"';"  ,(err, rows, fiels) => {  
-          //pool.query("INSERT INTO Patients (title,firstName,middleInitial,lastName,phone,email,sex,ssn,dateOfBirth,street,city,state,zip,insuranceCompany,plan,groupNumber,cardholder,medications,allergies,surgeries,familyHistory,addictions,questionnaire,symptoms,signature) VALUES ('"+patient.title+"','"+patient.firstName+"','"+patient.middleInitial+"','"+patient.lastName+"','"+patient.phone+"','"+patient.email+"','"+patient.sex+"','"+patient.ssn+"','"+patient.dateOfBirth+"','"+patient.street+"','"+patient.city+"','"+patient.state+"','"+patient.zip+"','"+patient.insuranceCompany+"','"+patient.plan+"','"+patient.groupNumber+"','"+patient.cardHolder+"','"+med+"','"+all+"','"+sur+"','"+fam+"','"+add+"','"+que+"','"+sym+"','"+patient.signature+"');" ,(err, rows, fiels) => {  
-         pool.query("UPDATE Patients SET title='"+patient.title+"',firstName='"+patient.firstName+"',middleInitial='"+patient.middleInitial+"',lastName='"+patient.lastName+"',phone='"+patient.phone+"',email='"+patient.email+"',sex='"+patient.sex+"',ssn='"+patient.ssn+"',dateOfBirth='"+patient.dateOfBirth+"',street='"+patient.street+"',city='"+patient.city+"',state='"+patient.state+"',zip='"+patient.zip+"',insuranceCompany='"+patient.insuranceCompany+"',plan='"+patient.plan+"',groupNumber='"+patient.groupNumber+"',cardHolder='"+patient.cardHolder+"',medications='"+med+"',allergies='"+ all+"',surgeries='"+sur+"',familyHistory='"+fam+"',addictions='"+add+"',questionnaire='"+que+"',symptoms='"+sym+"',signature='"+patient.signature+"' WHERE id = '"+patient.id+"';"  ,(err, rows, fiels) => {  
+          pool.query("UPDATE Patients SET title='"+patient.title+"',firstName='"+patient.firstName+"',middleInitial='"+patient.middleInitial+"',lastName='"+patient.lastName+"',phone='"+patient.phone+"',email='"+patient.email+"',sex='"+patient.sex+"',ssn='"+patient.ssn+"',dateOfBirth='"+patient.dateOfBirth+"',street='"+patient.street+"',city='"+patient.city+"',state='"+patient.state+"',zip='"+patient.zip+"',insuranceCompany='"+patient.insuranceCompany+"',plan='"+patient.plan+"',groupNumber='"+patient.groupNumber+"',cardHolder='"+patient.cardHolder+"',medications='"+med+"',allergies='"+ all+"',surgeries='"+sur+"',familyHistory='"+fam+"',addictions='"+add+"',questionnaire='"+que+"',symptoms='"+sym+"',signature='"+patient.signature+"' WHERE id = '"+patient.id+"';"  ,(err, rows, fiels) => {  
     
     if (!err) {
       res.json(rows);
@@ -86,8 +72,7 @@ router.put("/", async  (req, res) => {
       console.log(err);
     }
   });
-   //  res.status(201).send("Patient "+patient.firstName+" "+patient.lastName+" has beed added to the patient list.")
-    }
+     }
     else{
       res.status(400).send(result)
     }
@@ -97,6 +82,10 @@ router.put("/", async  (req, res) => {
 
 router.delete('/:id',(req,res,next)=> {
     var thisId = req.params;
+     if (typeof(thisId.id) == "undefined"){
+        res.status(404).send('No body in request.')
+     }
+     else {
    
        pool.query("Delete From Patients Where id ="+thisId.id+";",(err, rows, fiels) => {  
     if (!err) {
@@ -104,24 +93,20 @@ router.delete('/:id',(req,res,next)=> {
         res.status(400).send('Patient with ID: '+thisId.id+ ' not found.')
       }
       else{
-     // res.json(rows);}
-        res.status(200).send('Patient with ID: '+thisId.id+ ' has been deleted.')
+         res.status(200).send('Patient with ID: '+thisId.id+ ' has been deleted.')
      }
       console.log(rows.affectedRows);
     } else {
       
       console.log(err);
     }
-  });
-  //   res.status(201).send("Patient "+thisId.id+" has been deleted from patient list.")
+  });}
   
 });
 
 router.get('/record',async(req,res,next)=> {
-   // var contactId = req.params;
      var contact= req.body
-     console.log(contact)
-      console.log(typeof(req.body))
+     
      if (typeof(contact.id) == "undefined"){
         res.status(404).send('No body in request.')
      }
@@ -149,13 +134,11 @@ router.get('/record',async(req,res,next)=> {
       console.log(err);
     }
   });}}
-  //   res.status(201).send("Patient "+thisId.id+" has been deleted from contact list.")
   
 });
 
 router.get('/single/:id',(req,res,next)=> {
     var patId = req.params;
-    
    
        pool.query("select * From Patients Where id ="+patId.id+";",(err, rows, fiels) => {  
          
@@ -177,13 +160,12 @@ router.get('/single/:id',(req,res,next)=> {
       console.log(err);
     }
   });
-  //   res.status(201).send("Patient "+thisId.id+" has been deleted from patient list.")
-  
+      
 });
 
 router.get('/email/:id',(req,res,next)=> {
     var emailId = req.params;
-   
+     
        pool.query("select * From Patients Where email ='"+emailId.id+"';",(err, rows, fiels) => {  
          console.log(rows)
     if (rows<1)    {
@@ -203,12 +185,10 @@ router.get('/email/:id',(req,res,next)=> {
       console.log(err);
     }
   });
-  //   res.status(201).send("Patient "+thisId.id+" has been deleted from patient list.")
-  
+      
 });
  async function validatePatient(contact) {
-  //result="ok"
-  
+   
   if(validateTitle(contact.title)){
     return "Please use titles Dr, Mr, Mrs, Ms or Miss not "+ contact.title+"."
   }
@@ -247,9 +227,6 @@ router.get('/email/:id',(req,res,next)=> {
      if(validateDob(contact.dateOfBirth)){
    return "Please use only valid date of births not "+ contact.dateOfBirth+"."
    }
- //   if(validateLetters(contact.relationToPatient)){
- //  return "Please use only use letters in relation to patient not "+ contact.relationToPatient+"."
-//   }
     if(validateLetters(contact.signature)){
    return "Please use only use letters in signature not "+ contact.signature+"."
    }
@@ -291,10 +268,7 @@ function validateDob(dob) {
     if(dob.match(/(\d{4})-(\d{2})-(\d{2})/)){
      return false
    }
-   // if(!(/^(0?[1-9]|1[0-2])[\/](0?[1-9]|[1-2][0-9]|3[01])[\/]\d{4}$/.test(dob))){
-   //  return true
-  // }
-   return true
+      return true
 }
 function validateStreet(street) {
